Created single API class Extracted Arcana capability to API
