+++ src/api/java/divinerpg/api/package-info.java
@@ -1,4 +1,5 @@
-@API(apiVersion = Reference.API_VERSION, owner = Reference.MODID, provides = Reference.NAME + "|API")
+// TODO find out what 'provides' means
+@API(apiVersion = Reference.API_VERSION, owner = Reference.MODID, provides = "DivineAPI")
 package divinerpg.api;
 
 import net.minecraftforge.fml.common.API;
