+++ src/main/java/divinerpg/events/ArcanaTickHandler.java
@@ -1,8 +1,7 @@
 package divinerpg.events;
 
-import divinerpg.api.arcana.ArcanaAPI;
+import divinerpg.api.DivineAPI;
 import divinerpg.api.arcana.IArcana;
-import divinerpg.capabilities.ArcanaProvider;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.PlayerEvent;
@@ -14,7 +13,7 @@ public class ArcanaTickHandler {
 
     @SubscribeEvent
     public void onTick(PlayerTickEvent event) {
-        arcana = ArcanaAPI.getArcana(event.player);
+        arcana = DivineAPI.getArcana(event.player);
         if (event.phase == Phase.START) {
             // onTickStart(event.player);
         } else {
@@ -23,35 +22,35 @@ public class ArcanaTickHandler {
     }
 
     private void onTickEnd(EntityPlayer player) {
-        arcana = ArcanaAPI.getArcana(player);
+        arcana = DivineAPI.getArcana(player);
         arcana.regen(player);
     }
 
     private void onTickStart(EntityPlayer player) {
-        //arcana = ArcanaAPI.getArcana(player);
+        //arcana = DivineAPI.getArcana(player);
     }
 
     @SubscribeEvent
     public void onPlayerLoggedIn(PlayerEvent.PlayerLoggedInEvent event) {
-        arcana = ArcanaAPI.getArcana(event.player);
+        arcana = DivineAPI.getArcana(event.player);
         arcana.set(arcana.getMaxArcana());
     }
 
     @SubscribeEvent
     public void onPlayerLoggedOut(PlayerEvent.PlayerLoggedOutEvent event) {
-        arcana = ArcanaAPI.getArcana(event.player);
+        arcana = DivineAPI.getArcana(event.player);
         arcana.set(0);
     }
 
     @SubscribeEvent
     public void onPlayerRespawn(PlayerEvent.PlayerRespawnEvent event) {
-        arcana = ArcanaAPI.getArcana(event.player);
+        arcana = DivineAPI.getArcana(event.player);
         arcana.set(arcana.getMaxArcana());
     }
 
     @SubscribeEvent
     public void onPlayerChangedDimension(PlayerEvent.PlayerChangedDimensionEvent event) {
-        arcana = ArcanaAPI.getArcana(event.player);
+        arcana = DivineAPI.getArcana(event.player);
         arcana.set(arcana.getMaxArcana());
         event.player.addExperienceLevel(0);
     }
