+++ src/main/java/divinerpg/networking/message/MessageArcanaBar.java
@@ -1,6 +1,7 @@
 package divinerpg.networking.message;
 
 import divinerpg.DivineRPG;
+import divinerpg.api.DivineAPI;
 import divinerpg.api.arcana.IArcana;
 import divinerpg.client.ArcanaRenderer;
 import io.netty.buffer.ByteBuf;
@@ -9,7 +10,7 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 import net.minecraftforge.fml.relauncher.Side;
 
-import static divinerpg.capabilities.ArcanaProvider.ARCANA_CAP;
+import static divinerpg.api.arcana.ArcanaProvider.ARCANA_CAP;
 
 public class MessageArcanaBar implements IMessage {
 
@@ -48,7 +49,7 @@ public class MessageArcanaBar implements IMessage {
 		public IMessage onMessage(MessageArcanaBar msg, MessageContext ctx) {
 
 			if (ctx.side == Side.CLIENT) {
-				IArcana capability = DivineRPG.proxy.getPlayer().getCapability(ARCANA_CAP, null);
+				IArcana capability = DivineAPI.getArcana(DivineRPG.proxy.getPlayer());
 				capability.setMaxArcana(msg.max);
 				capability.setRegenDelay(msg.delay);
 				capability.set(msg.arcana);
