+++ src/main/java/divinerpg/objects/items/arcana/ItemArcaniumSaber.java
@@ -1,8 +1,7 @@
 package divinerpg.objects.items.arcana;
 
-import divinerpg.api.arcana.ArcanaAPI;
+import divinerpg.api.DivineAPI;
 import divinerpg.api.arcana.IArcana;
-import divinerpg.capabilities.ArcanaProvider;
 import divinerpg.objects.items.base.ItemModSword;
 import divinerpg.registry.ModSounds;
 import divinerpg.utils.TooltipLocalizer;
@@ -27,7 +26,7 @@ public class ItemArcaniumSaber extends ItemModSword {
 
     @Override
     public boolean onLeftClickEntity(ItemStack stack, EntityPlayer player, Entity entity) {
-        IArcana arcana = ArcanaAPI.getArcana(player);
+        IArcana arcana = DivineAPI.getArcana(player);
         if (arcana.getArcana() < 12) {
             return false;
         } else {
