+++ src/main/java/divinerpg/objects/items/arcana/ItemEnderScepter.java
@@ -1,8 +1,7 @@
 package divinerpg.objects.items.arcana;
 
-import divinerpg.api.arcana.ArcanaAPI;
+import divinerpg.api.DivineAPI;
 import divinerpg.api.arcana.IArcana;
-import divinerpg.capabilities.ArcanaProvider;
 import divinerpg.objects.items.base.ItemMod;
 import divinerpg.registry.DivineRPGTabs;
 import divinerpg.utils.PositionHelper;
@@ -34,7 +33,7 @@ public class ItemEnderScepter extends ItemMod {
     @Override
     public @Nonnull ActionResult<ItemStack> onItemRightClick(@Nonnull World world, @Nonnull EntityPlayer player,
             @Nonnull EnumHand hand) {
-        IArcana arcana = ArcanaAPI.getArcana(player);
+        IArcana arcana = DivineAPI.getArcana(player);
         if (!world.isRemote && arcana.getArcana() >= 75) {
             RayTraceResult pos = PositionHelper.rayTrace(player, 32, 1);
             int x = pos.getBlockPos().getX(), y = pos.getBlockPos().getY() + 1, z = pos.getBlockPos().getZ();
