+++ src/main/java/divinerpg/objects/items/arcana/ItemMeriksMissile.java
@@ -1,8 +1,7 @@
 package divinerpg.objects.items.arcana;
 
-import divinerpg.api.arcana.ArcanaAPI;
+import divinerpg.api.DivineAPI;
 import divinerpg.api.arcana.IArcana;
-import divinerpg.capabilities.ArcanaProvider;
 import divinerpg.objects.entities.entity.projectiles.EntityMerikMissile;
 import divinerpg.objects.items.base.ItemMod;
 import divinerpg.registry.DivineRPGTabs;
@@ -51,7 +50,7 @@ public class ItemMeriksMissile extends ItemMod {
         if (stack.getItem() == ModItems.meriksMissile && entityLiving instanceof EntityPlayer) {
             float charge = (float) (MAX_USE_DURATION - timeLeft) / 20.0F;
             EntityPlayer player = (EntityPlayer) entityLiving;
-            IArcana arcana = ArcanaAPI.getArcana(player);
+            IArcana arcana = DivineAPI.getArcana(player);
 
             if (charge > 1.0) {
                 charge = 1.0F;
