+++ src/main/java/divinerpg/objects/items/arcana/ItemShadowSaber.java
@@ -1,8 +1,7 @@
 package divinerpg.objects.items.arcana;
 
-import divinerpg.api.arcana.ArcanaAPI;
+import divinerpg.api.DivineAPI;
 import divinerpg.api.arcana.IArcana;
-import divinerpg.capabilities.ArcanaProvider;
 import divinerpg.objects.items.base.ItemModSword;
 import divinerpg.registry.ModSounds;
 import divinerpg.utils.TooltipHelper;
@@ -32,7 +31,7 @@ public class ItemShadowSaber extends ItemModSword {
     public boolean hitEntity(ItemStack stack, EntityLivingBase target, EntityLivingBase hitter) {
         if (hitter instanceof EntityPlayer) {
             EntityPlayer player = (EntityPlayer) hitter;
-            IArcana arcana = ArcanaAPI.getArcana(player);
+            IArcana arcana = DivineAPI.getArcana(player);
             if (arcana.getArcana() >= 12) {
                 player.playSound(ModSounds.SHADOW_SABER, 1, 1);
                 player.addPotionEffect(new PotionEffect(MobEffects.SPEED, 5 * 20, 1));
