+++ src/main/java/divinerpg/proxy/ClientProxy.java
@@ -1,10 +1,9 @@
 package divinerpg.proxy;
 
 import divinerpg.DivineRPG;
+import divinerpg.api.DivineAPI;
 import divinerpg.api.Reference;
-import divinerpg.api.arcana.ArcanaAPI;
 import divinerpg.api.arcana.IArcana;
-import divinerpg.capabilities.ArcanaProvider;
 import divinerpg.client.ArcanaRenderer;
 import divinerpg.client.ClientTicker;
 import divinerpg.enums.ParticleType;
@@ -197,7 +196,7 @@ public class ClientProxy extends CommonProxy {
         EntityPlayer player = DivineRPG.proxy.getPlayer();
 
         if (player != null) {
-            IArcana arcana = ArcanaAPI.getArcana(player);
+            IArcana arcana = DivineAPI.getArcana(player);
             if (arcana != null) {
                 arcana.set(amount);
             }
